requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.145
boto3-stubs==1.34.145
botocore==1.34.145
botocore-stubs==1.34.145
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langsmith==0.1.93
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
octoai==1.5.0
orjson==3.10.6
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
types-awscrt==0.21.2
types-boto3==1.0.2
types-s3transfer==0.10.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
